Added SimHash fuzzy file comparison.
Refactored to get rid of gosec complaints regarding the use of
Open and a potential integer overflow.

+++ internal/dfs/nocache_other.go
@@ -2,6 +2,6 @@
 
 package dfs
 
-func setNoCacheFD(_ int) error {
+func setNoCacheFD(_ uintptr) error {
 	return nil
 }
